updated readme and deleted input for privacy info

+++ code/make_template_json.py
@@ -64,11 +64,12 @@ def main():
         product_link = row["product_link"]
         product_name = row["product_name"]
         filename = os.path.join(PATH, img_folder_path, f"{index}.jpg")
-        img_path = download_image(product_link, filename)
+        img_path, status = download_image(product_link, filename)
         img_key = {"img_key": get_img_key(img_path)}
         cleaned_df.at[index, "img_key"] = img_key  # Add img_key to
 
-        os.remove(img_path) #delete image file after getting img_key
+        if status:
+            os.remove(img_path) #delete image file after getting img_key
 
     json_template = get_json(cleaned_df)
     print(json.dumps(json_template, indent=4, ensure_ascii=False))  # For testing purposes
